Reorg CCR
