+++ specification/ccr/follow_index_stats/FollowIndexStatsResponse.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { FollowIndexStats } from './FollowIndexStats'
+import { FollowIndexStats } from '@ccr/_types/FollowIndexStats'
 
 export class Response {
   body: { indices: FollowIndexStats[] }
