+++ specification/ccr/put_auto_follow_pattern/PutAutoFollowPatternRequest.ts
@@ -38,16 +38,26 @@ export interface Request extends RequestBase {
     remote_cluster: string
     follow_index_pattern?: IndexPattern
     leader_index_patterns?: IndexPatterns
-    max_outstanding_read_requests?: integer // default: 12
+    /** @server_default 12 */
+    max_outstanding_read_requests?: integer
     settings?: Dictionary<string, UserDefinedValue>
-    max_outstanding_write_requests?: integer // default: 9
-    read_poll_timeout?: Time // default: 1m
-    max_read_request_operation_count?: integer // default: 5120
-    max_read_request_size?: ByteSize // default: 32mb
-    max_retry_delay?: Time // default: 500ms
-    max_write_buffer_count?: integer // default: 2147483647
-    max_write_buffer_size?: ByteSize // default: 512mb
-    max_write_request_operation_count?: integer // default: 5120
-    max_write_request_size?: ByteSize // default: 9223372036854775807b
+    /** @server_default 9 */
+    max_outstanding_write_requests?: integer
+    /** @server_default 1m */
+    read_poll_timeout?: Time
+    /** @server_default 5120 */
+    max_read_request_operation_count?: integer
+    /** @server_default 32mb */
+    max_read_request_size?: ByteSize
+    /** @server_default 500ms */
+    max_retry_delay?: Time
+    /** @server_default 2147483647 */
+    max_write_buffer_count?: integer
+    /** @server_default 512mb */
+    max_write_buffer_size?: ByteSize
+    /** @server_default 5120 */
+    max_write_request_operation_count?: integer
+    /** @server_default 9223372036854775807b */
+    max_write_request_size?: ByteSize
   }
 }
