+++ specification/ccr/resume_follow_index/ResumeFollowIndexRequest.ts
@@ -31,7 +31,6 @@ export interface Request extends RequestBase {
   path_parts?: {
     index: IndexName
   }
-  query_parameters?: {}
   body?: {
     max_outstanding_read_requests?: long
     max_outstanding_write_requests?: long
