+++ specification/ccr/stats/CcrStatsRequest.ts
@@ -24,7 +24,4 @@ import { RequestBase } from '@_types/Base'
  * @since 6.5.0
  * @stability TODO
  */
-export interface Request extends RequestBase {
-  query_parameters?: {}
-  body?: {}
-}
+export interface Request extends RequestBase {}
